[TD-10901]<fix>: Solve the conversion and serialization problems caused by long SQL.

+++ src/inc/taosmsg.h
@@ -486,7 +486,7 @@ typedef struct {
   SInterval   interval;
   SSessionWindow sw;            // session window
   uint16_t    tagCondLen;       // tag length in current query
-  uint16_t    colCondLen;       // column length in current query
+  int32_t    colCondLen;       // column length in current query
   int16_t     numOfGroupCols;   // num of group by columns
   int16_t     orderByIdx;
   int16_t     orderType;        // used in group by xx order by xxx
